+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.11.1 under development
 
 - New #177: Add `hasTopicsEnabled` field to `User` type.
+- New #177: Add `SendMessageDraft` method.
 
 ## 0.11.0 December 29, 2025